+++ contingency/contingency-api/src/main/java/com/powsybl/contingency/contingency/list/AbstractEquipmentCriterionContingencyList.java
@@ -7,11 +7,15 @@
 package com.powsybl.contingency.contingency.list;
 
 import com.google.common.collect.ImmutableList;
+import com.powsybl.contingency.Contingency;
+import com.powsybl.contingency.ContingencyElement;
 import com.powsybl.contingency.contingency.list.criterion.*;
 import com.powsybl.iidm.network.IdentifiableType;
+import com.powsybl.iidm.network.Network;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 /**
  * @author Etienne Lesot <etienne.lesot@rte-france.com>
@@ -21,10 +25,16 @@ public abstract class AbstractEquipmentCriterionContingencyList implements Conti
     private final IdentifiableType identifiableType;
     private final List<PropertyCriterion> propertyCriteria;
     private final RegexCriterion regexCriterion;
+    private final Criterion countryCriterion;
+    private final Criterion nominalVoltageCriterion;
 
-    protected AbstractEquipmentCriterionContingencyList(String name, IdentifiableType identifiableType, List<PropertyCriterion> propertyCriteria, RegexCriterion regexCriterion) {
+    protected AbstractEquipmentCriterionContingencyList(String name, IdentifiableType identifiableType,
+                                                        Criterion countryCriterion, Criterion nominalVoltageCriterion,
+                                                        List<PropertyCriterion> propertyCriteria, RegexCriterion regexCriterion) {
         this.name = Objects.requireNonNull(name);
         this.identifiableType = identifiableType;
+        this.countryCriterion = countryCriterion;
+        this.nominalVoltageCriterion = nominalVoltageCriterion;
         this.propertyCriteria = ImmutableList.copyOf(Objects.requireNonNull(propertyCriteria));
         this.regexCriterion = regexCriterion;
     }
@@ -32,6 +42,17 @@ public abstract class AbstractEquipmentCriterionContingencyList implements Conti
     @Override
     public String getName() {
         return name;
+    }
+
+    @Override
+    public List<Contingency> getContingencies(Network network) {
+        return network.getIdentifiableStream(getIdentifiableType())
+                .filter(identifiable -> countryCriterion == null || countryCriterion.filter(identifiable, getIdentifiableType()))
+                .filter(identifiable -> nominalVoltageCriterion == null || nominalVoltageCriterion.filter(identifiable, getIdentifiableType()))
+                .filter(identifiable -> getPropertyCriteria().stream().allMatch(propertyCriterion -> propertyCriterion.filter(identifiable, getIdentifiableType())))
+                .filter(identifiable -> getRegexCriterion() == null || getRegexCriterion().filter(identifiable, getIdentifiableType()))
+                .map(identifiable -> new Contingency(identifiable.getId(), ContingencyElement.of(identifiable)))
+                .collect(Collectors.toList());
     }
 
     public IdentifiableType getIdentifiableType() {
@@ -46,7 +67,11 @@ public abstract class AbstractEquipmentCriterionContingencyList implements Conti
         return regexCriterion;
     }
 
-    public abstract Criterion getCountryCriterion();
+    public Criterion getCountryCriterion() {
+        return countryCriterion;
+    }
 
-    public abstract Criterion getNominalVoltageCriterion();
+    public Criterion getNominalVoltageCriterion() {
+        return nominalVoltageCriterion;
+    }
 }
