+++ contingency/contingency-api/src/main/java/com/powsybl/contingency/contingency/list/InjectionCriterionContingencyList.java
@@ -6,25 +6,18 @@
  */
 package com.powsybl.contingency.contingency.list;
 
-import com.powsybl.contingency.Contingency;
-import com.powsybl.contingency.ContingencyElement;
 import com.powsybl.contingency.contingency.list.criterion.PropertyCriterion;
 import com.powsybl.contingency.contingency.list.criterion.RegexCriterion;
 import com.powsybl.contingency.contingency.list.criterion.SingleCountryCriterion;
 import com.powsybl.contingency.contingency.list.criterion.SingleNominalVoltageCriterion;
 import com.powsybl.iidm.network.IdentifiableType;
-import com.powsybl.iidm.network.Network;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * @author Etienne Lesot <etienne.lesot@rte-france.com>
  */
 public class InjectionCriterionContingencyList extends AbstractEquipmentCriterionContingencyList {
-
-    private final SingleCountryCriterion singleCountryCriterion;
-    private final SingleNominalVoltageCriterion singleNominalVoltageCriterion;
 
     public InjectionCriterionContingencyList(String name, String identifiableType,
                                              SingleCountryCriterion singleCountryCriterion,
@@ -35,34 +28,11 @@ public class InjectionCriterionContingencyList extends AbstractEquipmentCriterio
     public InjectionCriterionContingencyList(String name, IdentifiableType identifiableType,
                                              SingleCountryCriterion singleCountryCriterion,
                                              SingleNominalVoltageCriterion singleNominalVoltageCriterion, List<PropertyCriterion> propertyCriteria, RegexCriterion regexCriterion) {
-        super(name, identifiableType, propertyCriteria, regexCriterion);
-        this.singleCountryCriterion = singleCountryCriterion;
-        this.singleNominalVoltageCriterion = singleNominalVoltageCriterion;
+        super(name, identifiableType, singleCountryCriterion, singleNominalVoltageCriterion, propertyCriteria, regexCriterion);
     }
 
     @Override
     public String getType() {
         return "injectionCriterion";
-    }
-
-    @Override
-    public List<Contingency> getContingencies(Network network) {
-        return network.getIdentifiableStream(getIdentifiableType())
-                .filter(identifiable -> singleCountryCriterion == null || singleCountryCriterion.filter(identifiable, getIdentifiableType()))
-                .filter(identifiable -> singleNominalVoltageCriterion == null || singleNominalVoltageCriterion.filter(identifiable, getIdentifiableType()))
-                .filter(identifiable -> getPropertyCriteria().stream().allMatch(propertyCriterion -> propertyCriterion.filter(identifiable, getIdentifiableType())))
-                .filter(identifiable -> getRegexCriterion() == null || getRegexCriterion().filter(identifiable, getIdentifiableType()))
-                .map(identifiable -> new Contingency(identifiable.getId(), ContingencyElement.of(identifiable)))
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public SingleCountryCriterion getCountryCriterion() {
-        return singleCountryCriterion;
-    }
-
-    @Override
-    public SingleNominalVoltageCriterion getNominalVoltageCriterion() {
-        return singleNominalVoltageCriterion;
     }
 }
