+++ contingency/contingency-api/src/main/java/com/powsybl/contingency/contingency/list/ThreeWindingsTransformerCriterionContingencyList.java
@@ -6,58 +6,28 @@
  */
 package com.powsybl.contingency.contingency.list;
 
-import com.powsybl.contingency.Contingency;
-import com.powsybl.contingency.ContingencyElement;
 import com.powsybl.contingency.contingency.list.criterion.PropertyCriterion;
 import com.powsybl.contingency.contingency.list.criterion.RegexCriterion;
 import com.powsybl.contingency.contingency.list.criterion.SingleCountryCriterion;
 import com.powsybl.contingency.contingency.list.criterion.ThreeNominalVoltageCriterion;
 import com.powsybl.iidm.network.IdentifiableType;
-import com.powsybl.iidm.network.Network;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * @author Etienne Lesot <etienne.lesot@rte-france.com>
  */
 public class ThreeWindingsTransformerCriterionContingencyList extends AbstractEquipmentCriterionContingencyList {
 
-    private final SingleCountryCriterion singleCountryCriterion;
-    private final ThreeNominalVoltageCriterion threeNominalVoltageCriterion;
-
     public ThreeWindingsTransformerCriterionContingencyList(String name,
                                                             SingleCountryCriterion singleCountryCriterion,
                                                             ThreeNominalVoltageCriterion threeNominalVoltageCriterion,
                                                             List<PropertyCriterion> propertyCriteria, RegexCriterion regexCriterion) {
-        super(name, IdentifiableType.THREE_WINDINGS_TRANSFORMER, propertyCriteria, regexCriterion);
-        this.singleCountryCriterion = singleCountryCriterion;
-        this.threeNominalVoltageCriterion = threeNominalVoltageCriterion;
+        super(name, IdentifiableType.THREE_WINDINGS_TRANSFORMER, singleCountryCriterion, threeNominalVoltageCriterion, propertyCriteria, regexCriterion);
     }
 
     @Override
     public String getType() {
         return "threeWindingsTransformerCriterion";
-    }
-
-    @Override
-    public List<Contingency> getContingencies(Network network) {
-        return network.getIdentifiableStream(getIdentifiableType())
-                .filter(identifiable -> singleCountryCriterion == null || singleCountryCriterion.filter(identifiable, getIdentifiableType()))
-                .filter(identifiable -> threeNominalVoltageCriterion == null || threeNominalVoltageCriterion.filter(identifiable, getIdentifiableType()))
-                .filter(identifiable -> getPropertyCriteria().stream().allMatch(propertyCriterion -> propertyCriterion.filter(identifiable, getIdentifiableType())))
-                .filter(identifiable -> getRegexCriterion() == null || getRegexCriterion().filter(identifiable, getIdentifiableType()))
-                .map(identifiable -> new Contingency(identifiable.getId(), ContingencyElement.of(identifiable)))
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public SingleCountryCriterion getCountryCriterion() {
-        return singleCountryCriterion;
-    }
-
-    @Override
-    public ThreeNominalVoltageCriterion getNominalVoltageCriterion() {
-        return threeNominalVoltageCriterion;
     }
 }
